+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Pre-Release 1.0.0-beta (XX XXX, 202X)
+## Pre-Release 1.0.0-beta (30 December, 2024)
 * Minor bug fixes
 * Moved the lockfile creation to the exclude folder to avoid write premission 
 in script path